% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{branch}
\alias{branch}
\title{Branch a file to work on locally as a method for version control.}
\usage{
branch(file, inFolder = NULL)
}
\arguments{
\item{file}{Master file to create a branch for}

\item{inFolder}{An identifer to narrow the search in case there are multiple files with same name but in different folders (i.e. "Codes/Model1").}
}
\value{
No return value.
}
\description{
Clones a file to the Branches folder to work on locally as a version control method. The function creates
a subfolder in Branches with the username on the users computer.
}
\examples{
branch("Project Master.R")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
