% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{build_cabinet}
\alias{build_cabinet}
\title{Builds the file cabinet}
\usage{
build_cabinet()
}
\value{
No return value
}
\description{
Searches through all folders in the project directory for all files to store in
the cabinet. All paths are stored relative to the projects working directory and stored in a package environment variable called cabinet.
}
\examples{
build_cabinet()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
