% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{build_query}
\alias{build_query}
\title{Build a query for Google BigQuery from a text string}
\usage{
build_query(query, standard = T, limit = NULL, show = F)
}
\arguments{
\item{query}{A character string written in SQL language.}

\item{standard}{Appends "#standardSQL" to the top if set to T. Default is F.}

\item{limit}{Appends "LIMIT = 1000", for example, or whatever integer value limit is set to. Default is NULL and will not append anything.}

\item{show}{A boolean (T, F) indicator of whether to print the query to the console for visualization.}
}
\value{
A character string.
}
\description{
Takes a character string and concatenates some optional lines
}
\examples{
query = "SELECT * FROM table"
build_query(query, standard = T, limit = 1000, show = T)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
