% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{clone}
\alias{clone}
\title{Clone the project into a branch directory denoted by the username as a method for version control.}
\usage{
clone(ignore = c("RData", "csv", "xls", "xlsx"), overwrite = T)
}
\arguments{
\item{ignore}{file extensions to not copy. Default is c("RData", "csv", "xls", "xlsx").}

\item{overwrite}{Boolean (T, F) of whether to overwrite existing copies}
}
\value{
No return value.
}
\description{
Clones the project into the Branches folder to work on locally as a version control method. The function creates
a subfolder in Branches with the username on the users computer.
}
\examples{
clone()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
