% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\docType{data}
\name{env_loc}
\alias{env_loc}
\title{Creates a project environment variable}
\format{An object of class \code{environment} of length 57.}
\usage{
env_loc
}
\description{
An environment variable
}
\examples{
names(proj.env)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
\keyword{datasets}
