% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{exit_proj}
\alias{exit_proj}
\title{Exit a project packrat mode}
\usage{
exit_proj()
}
\value{
No return value
}
\description{
Wrapper for packrat's disable() function
}
\examples{
exit_proj()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
