% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{find_dirs}
\alias{find_dirs}
\title{Find the project root directory}
\usage{
find_dirs()
}
\value{
List containing root.dir, current.dir, and found_wd
}
\description{
Finds important project directories
}
\examples{
finds_dirs()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
