% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{get_file_path}
\alias{get_file_path}
\title{Get a full file path}
\usage{
get_file_path(file, inFolder = NULL, recall = T, allowMult = F)
}
\arguments{
\item{file}{A character string giving the name of the file to get the full folder path for (i.e. "Project Master.R").}

\item{inFolder}{An identifer to narrow the search in case there are multiple files with same name but in different folders (i.e. "Codes/Model1").}

\item{recall}{A boolean (T, F) indicator specifying whether to rebuild the cabinet if the file was not found the first time. The default is T.}

\item{allowMult}{A boolean (T, F) indicator specifying whether to allow multiple file paths upon return. The default is F.}
}
\value{
A character string giving the full file path to the specified file.
}
\description{
Returns the full file path to the file. It searches in the package environment variable cabinet for the path.
}
\examples{
get_file_path("Model1.R")
get_file_path("Model.R", inFolder = "Codes")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
