% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{get_output_dir}
\alias{get_output_dir}
\title{Get the desired output directory}
\usage{
get_output_dir(doc = F, file = NULL, inFolder = NULL)
}
\arguments{
\item{doc}{Boolean (T, F) indicator of whether to set the output directory to the "Output" folder (used for R scripts calling the function) or the "Documentation" folder (used for Rmd scripts calling the function).}
}
\value{
A character string giving a full folder path for the desired output directory.
}
\description{
Gets a standard output directory path to be used for saving data and graphical objects based on the folder path of the R script the function is used in.
}
\details{
The output directory will be selected to mimic the structure of the Codes folder (created at the set up of the project) and any subfolders created afterwords.
If the R script exists in "Users/alexhubbard/Project/Codes/Model1" then the output directory will be "Users/alexhubbard/Project/Output/Model1". If an script is insted
an Rmd file, the output directory will change to "Users/alexhubbard/Project/Documentation/Model1".
}
\examples{
link_file()
get_output_dir()
get_output_dir(doc = T)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
