% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{get_proj_cur_dir}
\alias{get_proj_cur_dir}
\title{Get the current directory of the project}
\usage{
get_proj_cur_dir()
}
\value{
No return value
}
\description{
Updates the project environment with the current directory
}
\examples{
get_proj_cur_dir()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
