% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{get_proj_env}
\alias{get_proj_env}
\title{Loads the project environment variable}
\usage{
get_proj_env(...)
}
\description{
An environment variable
}
\examples{
proj.env = get_proj_env()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
