% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{ggplot_grid}
\alias{ggplot_grid}
\title{Grid of ggplot objects}
\usage{
ggplot_grid(g, ...)
}
\arguments{
\item{g}{A single ggplot object or a list of ggplot object defined as list(g1, g2, ...).}

\item{...}{Other arguements passed to grid.arrange.}
}
\value{
A plot
}
\description{
Modifies grid.arrange from the gridExtra package to take a list of objects of ggplot objects.
}
\examples{
library(ggplot2)
g1 = ggplot(iris) +
  ggtitle("Title", subtitle = "Subtitle") +
  geom_point(aes(x = Sepal.Width, y = Sepal.Length, color = Species), size = 4) +
  od_theme()

g2 = ggplot(iris) +
  ggtitle("Title", subtitle = "Subtitle") +
  geom_point(aes(x = Sepal.Width, y = Sepal.Length, color = Species), size = 4) +
  od_theme(n = length(unique(iris$Species)))

ggplot_grid(list(g1, g2), nrow = 1, ncol = 2)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
