% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{ggsave2}
\alias{ggsave2}
\title{A modified ggsave function}
\usage{
ggsave2(filename, plot = last_plot(), device = NULL, path = NULL,
  combine = F, scale = 1, width = NA, height = NA, units = c("in",
  "cm", "mm"), dpi = 300, limitsize = TRUE, ...)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot or plots (in a list) to save, defaults to last plot displayed.}

\item{device}{Device to use. Can be either be a device function (e.g. png), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{Path to save plot to (combined with filename).}

\item{combine}{Boolean (T, F) indicator of whether or not to combine the output from a list of ggplot items passed in plot to a single pdf.}

\item{...}{Other arguments passed on to graphics device.}
}
\value{
No return value
}
\description{
Redefined ggsave function from ggplot2 allowing to combine a list of ggplot items into a single pdf output. This function is used in the save function defined in this package.
}
\examples{
See ggplot2 package for examples.
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
