% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{git_branch}
\alias{git_branch}
\title{Create a branch using git}
\usage{
git_branch(branch = NULL)
}
\arguments{
\item{branch}{Branch name}
}
\value{
No return value
}
\description{
Creates a git branch. If branch left NULL then uses the current username on the computer.
The function also sets the newly created branch to be upstream from the master branch.
}
\examples{
git_branch()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
