% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{git_clone}
\alias{git_clone}
\title{Clone a git repository}
\usage{
git_clone(repo = NULL, directory = NULL)
}
\arguments{
\item{repo}{Character string path to the git repository to clone}

\item{directory}{Character string path where to clone the git repository to}
}
\value{
No return value
}
\description{
Clones a git repository
}
\examples{
clone(repo = "/users/username/Documents/repo", directory = "/users/username/Documents/clones/repo")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
