% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{git_commit}
\alias{git_commit}
\title{Commit changes to the current branch}
\usage{
git_commit(message)
}
\arguments{
\item{message}{A character string to provide with git commit}
}
\value{
No return value
}
\description{
Interactive of git merge a file from 2 branches
}
\examples{
git_commit("init")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
