% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{git_diff}
\alias{git_diff}
\title{Git diff}
\usage{
git_diff(branch1 = NULL, branch2 = NULL, file = NULL, inFolder = NULL,
  mode = "sidebyside")
}
\arguments{
\item{branch1}{The first branch to compare}

\item{branch2}{The second branch to compare}

\item{file}{The file name with extension to compare}

\item{inFolder}{An identifer to narrow the search in case there are multiple files with same name but in different folders (i.e. "Codes/Model1").}

\item{mode}{A character string of the display mode used by diffobj}
}
\value{
No return value
}
\description{
Git diff a file between two branches using diffobj. If branch2 is left NULL then uses the current branch.
}
\examples{
git_diff(branch1 = "master", branch2 = "alexhubbard", file = "Example File.R")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
