% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{git_merge}
\alias{git_merge}
\title{Merge two branches using git or interactive method}
\usage{
git_merge(with = NULL, this = NULL, file = NULL, inFolder = NULL)
}
\arguments{
\item{with}{Branch name want to merge current branch with}

\item{this}{Current branch}

\item{file}{File name and extenstion to merge with}

\item{inFolder}{An identifer to narrow the search in case there are multiple files with same name but in different folders (i.e. "Codes/Model1").}

\item{interactive}{Boolean (T, F) indicator of whether to interactively merge (T) or git merge (F)}
}
\value{
No return value
}
\description{
Interactive of git merge a file from 2 branches
}
\examples{
git_merge(with = "master")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
