% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{git_pull}
\alias{git_pull}
\title{Git pull}
\usage{
git_pull()
}
\value{
No return value
}
\description{
Git pull the master branch into the current branch.
}
\examples{
git_pull()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
