% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{git_push}
\alias{git_push}
\title{Git push}
\usage{
git_push()
}
\value{
No return value
}
\description{
Git push the current branch into the master branch.
}
\examples{
git_push()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
