% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{install.packges}
\alias{install.packges}
\title{This overwrites the base install.packages function to only install the function in the user's project library
for package version control}
\usage{
install.packges(..., lib = .libPaths()[1])
}
\arguments{
\item{...}{First, package name, then other named arguments passed to the base install.packages function}

\item{lib}{Path to users package library}
}
\value{
Normally library returns (invisibly) the list of attached packages, but TRUE or FALSE if logical.return is TRUE.
When called as library() it returns an object of class "libraryIQR", and for library(help=), one of class "packageInfo".
}
\description{
This overwrites the base install.packages function to only install the function in the user's project library
for package version control
}
\examples{
install.packages("packageName")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
