% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{install_req_packages}
\alias{install_req_packages}
\title{Install required packages}
\usage{
install_req_packages()
}
\value{
No return value
}
\description{
Installs required packages stored in ./Function/required_packages.csv
}
\examples{
install_req_packages()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
