% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{library}
\alias{library}
\title{Redefined library, require, and install.packages functions to only look in the project library
This overwrites the base library function to only look in the user's project library to load a package}
\usage{
library(..., lib.loc = NULL)
}
\arguments{
\item{...}{First, package name, then other named arguments passed to the base library function}

\item{lib.loc}{Path to users package library}
}
\value{
No return value
}
\description{
for package version control
}
\examples{
library(projectmap)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
