% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{lock}
\alias{lock}
\title{Lock all the project variables}
\usage{
lock(proj.env)
}
\value{
No return value
}
\description{
Lock all project environment variablers
}
\examples{
proj.env = lock(proj.env)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
