% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\docType{data}
\name{od.colors}
\alias{od.colors}
\title{Creates the Opendoor color scheme}
\format{An object of class \code{character} of length 11.}
\usage{
od.colors
}
\description{
A vector of named hexidecimal colors
}
\examples{
od.colors
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
\keyword{datasets}
