% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{od_theme}
\alias{od_theme}
\title{Opendoor's ggplot2 theme}
\usage{
od_theme(palette = "main", discrete = T, reverse = F, addblack = F,
  n = NULL, colors = NULL)
}
\arguments{
\item{palette}{A character string giving the name of one of the palattes ("main", "cool", "warm", "grey").}

\item{discrete}{A boolean(T, F) indicator setting whether to use a discrete or continuous theme. The default is F.}

\item{reverse}{A boolean(T, F) indicator for whether to reverse the color palette. The default is F.}

\item{addblack}{A boolean(T, F) indicator if want to add black to the color palatte. The default is F.}

\item{n}{An optional numeric value specifying the number of discrete colors to include in the palette.}

\item{colors}{An optional character vector of colors to use.}
}
\value{
A ggplot theme object
}
\description{
The full list of colors of the palette are: blue = "#1c85e8", navy = "#1d2c4c",
iris = "#ab80de", turquoise = "#68e1c7", citrine = "#ffc259", ruby = "#ff7f82",
lightgrey = "#babcbc", bluegrey = "#7b9cb1", coolgrey = "#506d7e", warmgrey = "#525975",
lightgreytint = "#f0f0f0".
}
\examples{
library(ggplot2)
ggplot(iris) +
  ggtitle("Title", subtitle = "Subtitle") +
  geom_point(aes(x = Sepal.Width, y = Sepal.Length, color = Species), size = 4) +
  od_theme()

ggplot(iris) +
  ggtitle("Title", subtitle = "Subtitle") +
  geom_point(aes(x = Sepal.Width, y = Sepal.Length, color = Species), size = 4) +
  od_theme(n = length(unique(iris$Species)))

ggplot(iris) +
  ggtitle("Title", subtitle = "Subtitle") +
  geom_point(aes(x = Sepal.Width, y = Sepal.Length, color = Species), size = 4) +
  od_theme(colors = c("turquoise", "citrine", "iris"))
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
