% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\docType{data}
\name{proj.env}
\alias{proj.env}
\title{Creates a project environment variable}
\format{An object of class \code{environment} of length 3.}
\usage{
proj.env
}
\description{
An environment variable
}
\examples{
names(proj.env)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
\keyword{datasets}
