% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{push_merge}
\alias{push_merge}
\title{Push merged banch with the master origin file as a method for version control.}
\usage{
push_merge(file, inFolder = NULL, user = NULL, open = F)
}
\arguments{
\item{file}{Master file to merge branch with.}

\item{inFolder}{An identifer to narrow the search in case there are multiple files with same name but in different folders (i.e. "Codes/Model1").}

\item{user}{A text string giving the username subfolder in the Branches folder to look for the cloned master file.}

\item{open}{Boolean (T, F) to open the file after branching}
}
\value{
No return value.
}
\description{
Merges a clone of the master with the master origin file.
}
\examples{
merge_branch("Project Master.R")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
