% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{read_file}
\alias{read_file}
\title{Read in a file in a standardized way}
\usage{
read_file(file, inFolder = NULL, showProgress = F, na.strings = c("NULL",
  "NA", "na", "N/A", "n/a", "<NA>", "NONE", "-", ".", "", " ", "NaN", "nan",
  "Inf", "-Inf"), envir = .GlobalEnv, ...)
}
\arguments{
\item{file}{A character string giving the name of the file to get the full folder path for (i.e. "Project Master.R").}

\item{inFolder}{An identifer to narrow the search in case there are multiple files with same name but in different folders (i.e. "Codes/Model1").}

\item{showProgress}{A boolean (T, F) indicator specifying whether to show the read in progress if using data.table's fread.}

\item{na.strings}{A vector of character strings to convert to NA}

\item{envir}{The environment to load the data to}

\item{...}{Other arguments to pass to data.tables fread, the base load or readRDS, or xlsx's read.xlsx.}
}
\value{
A data object (data.table or data.frame).
}
\description{
A wrapper function to read in a file containing data. It uses the file extenstion to determine whether to
use the base load or readRDS function for RData and rds files, data.table's fread function for csv files, or xlsx's read.xlsx for xls and xlsx files.
}
\examples{
read_file("Model1.R")
read_file("Model1.R", inFolder = "Codes")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
