% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{remove.packages}
\alias{remove.packages}
\title{This overwrites the base remove.packages function to only remove the function in the user's project library
for package version control}
\usage{
remove.packages(..., lib = NULL)
}
\arguments{
\item{...}{First, package name, then other named arguments passed to the base install.packages function}

\item{lib}{Path to users package library}
}
\value{
No return value
}
\description{
This overwrites the base remove.packages function to only remove the function in the user's project library
for package version control
}
\examples{
remove.packages("packageName")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
