% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{remove_file}
\alias{remove_file}
\title{Removes a file from the project directory and the file cabinet}
\usage{
remove_file(files)
}
\arguments{
\item{files}{A vector of character strings containing full file
paths of files to removed. The files will be deleted from the actual
folders on memory and in the project environment's cabinet.}
}
\value{
No return value
}
\description{
Removes files from the project directory and the file cabinet.
}
\examples{
remove_file(c("./Codes/Model1.R", "./Codes/Model1.R"))
remove_file(get_file_path("Model1.R"))
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
