% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{render_file}
\alias{render_file}
\title{Render an Rmd}
\usage{
render_file(file, inFolder = NULL, quiet = T, clean = T, ...)
}
\arguments{
\item{file}{A character string giving the name of the file, including the extension, to be saved.}

\item{inFolder}{An identifer to narrow the search in case there are multiple files with same name but in different folders (i.e. "Codes/Model1").}

\item{...}{Other parameters pased to rmarkdown::render}
}
\value{
No return value.
}
\description{
The functions uses knitr and rmarkdown to render an Rmd file and save it to the Documentation folder.
}
\examples{
render_file("Example.Rmd", inFolder = NULL, ...)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
