% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{require}
\alias{require}
\title{This overwrites the base require function to only look in the user's project library to load a package
for package version control}
\usage{
require(..., lib.loc = proj.env$libPath)
}
\arguments{
\item{...}{First, package name, then other named arguments passed to the base require function}

\item{lib.loc}{Path to users package library}
}
\value{
Normally library returns (invisibly) the list of attached packages, but TRUE or FALSE if logical.return is TRUE.
When called as library() it returns an object of class "libraryIQR", and for library(help=), one of class "packageInfo".
}
\description{
This overwrites the base require function to only look in the user's project library to load a package
for package version control
}
\examples{
require(projectmap)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
