% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{reset_init}
\alias{reset_init}
\title{Reset initializations tate}
\usage{
reset_init()
}
\arguments{
\item{build}{Boolean (T, F) indicator of whether to rebuild the cabinet or not. Default is F.}

\item{newroot}{Boolean (T, F) indicator of whether to reset root.dir}
}
\value{
No return value
}
\description{
Resets the project environment to its default values.
}
\examples{
reset_init()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
