% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{reset_proj_env}
\alias{reset_proj_env}
\title{Reset the project environment}
\usage{
reset_proj_env(build = F, newroot = F)
}
\arguments{
\item{build}{Boolean (T, F) indicator of whether to rebuild the cabinet or not. Default is F.}

\item{newroot}{Boolean (T, F) indicator of whether to reset root.dir}
}
\value{
No return value
}
\description{
Resets the project environment to its default values.
}
\examples{
reset_proj_env()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
