% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{run_proj_model}
\alias{run_proj_model}
\title{Return boolean of project model to be executed from the Project Master.R file.}
\usage{
run_proj_model(model)
}
\arguments{
\item{...}{The names and boolean values assigned to each model (i.e. set_proj_models(Model1 = T, model2 = F, ...))}
}
\value{
Boolean (T,F) indicator
}
\description{
Returns a boolean (T, F) indicator used as a flag telling "Project Master.R" whether to run a group of R scripts.
}
\examples{
if(run_proj_model("Model1")){
  source_file("Model1.R", inFolder = "Codes")
}
)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
