% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{save_proj_env}
\alias{save_proj_env}
\title{Save the project environment variable}
\usage{
save_proj_env(proj.env)
}
\description{
An environment variable
}
\examples{
save_proj_env(proj.env)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
