% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{set_proj_env}
\alias{set_proj_env}
\title{Updates the project environment}
\usage{
set_proj_env(...)
}
\arguments{
\item{...}{Named objects to assign to the project environment variable}
}
\description{
An environment variable
}
\examples{
set_proj_env(test = "test")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
