% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{set_proj_lib}
\alias{set_proj_lib}
\title{Set the path to the project library}
\usage{
set_proj_lib(init = F, app = F)
}
\arguments{
\item{init}{Boolean (T, F) to save the original library path upon project initiation}

\item{app}{Boolean (T, F) indicator to tell the function that it is being executed from within the app directory}
}
\value{
No return value
}
\description{
Sets the path to the project library
}
\examples{
set_proj_lib()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
