% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{set_proj_models}
\alias{set_proj_models}
\title{Set the project models to be executed from the "Project Master.R" file}
\usage{
set_proj_models(...)
}
\arguments{
\item{...}{The names and boolean values assigned to each model (i.e. set_proj_models(Model1 = T, model2 = F, ...))}
}
\value{
No return value
}
\description{
Assign a name and boolean (T, F) indicator to the project models you want to run when you source "Project Master.R".
}
\examples{
link_to_proj()
set_proj_models(
   Model1 = T,
   model2 = F
)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
