% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{source_file}
\alias{source_file}
\title{Source a file for the main.R file execution.}
\usage{
source_file(file, inFolder = NULL, docname = NULL,
  dont_unload = NULL, ...)
}
\arguments{
\item{file}{A character string giving the name of the file to get the full folder path for (i.e. "main.R").}

\item{inFolder}{An identifer to narrow the search in case there are multiple files with same name but in different folders (i.e. "Codes/Model1").}

\item{dont_unload}{A character list of packages names to prevent from being unloaded}

\item{...}{Additional parameters to pass to source}
}
\value{
No return value
}
\description{
A wrapper function for the base source command but also perfoms some backend functions to track the progress of the
files executed in "main.R" as well as updating the project progress bar. This function should only be used in the "main.R" script.
}
\examples{
link_to_proj()
set_proj_models(
  Model1 = T,
  Model2 = T
)
if(execute_proj_model("Model1")){
  source_file("Model1.R", inFolder = "Codes")
}
if(execute_proj_model("Model2")){
  source_file("Model2.R", inFolder = "Codes")
}
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
