% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{sum_dt}
\alias{sum_dt}
\title{A modified sum function}
\usage{
sum_dt(x, na.rm = F)
}
\arguments{
\item{x}{A vector of numeric values}

\item{na.rm}{A boolean (T, F) indicator of whether to include NA values or not in the calculation.}
}
\value{
A numeric value
}
\description{
Corrects the problem of returning 0 when adding up a vector of all NA's when na.rm = T
}
\examples{
sum_dt(c(1, 2, NA, 3), na.rm = T)
DT[, sum(col, na.rm = T)]
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
