% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{to_na}
\alias{to_na}
\title{Convert NaN and Inf to NA}
\usage{
to_na(DT)
}
\arguments{
\item{DT}{A data.table object}
}
\value{
A data object (data.table or data.frame).
}
\description{
Converts NaN, Inf, and -Inf values in a vecot to NA
#' @examples
data = data.table(x = c(1, 2, Inf, NaN, NA), y = c(1, 2, 3, 4, 5))
data = to_na(data)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
