% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{unload.packages}
\alias{unload.packages}
\title{This function unloads packages}
\usage{
unload.packages(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package names}
}
\value{
No return value
}
\description{
This function unloads packages
}
\examples{
unload.packages("packageName")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
