% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{unlock}
\alias{unlock}
\title{Unlock all the project variables}
\usage{
unlock(proj.env)
}
\value{
No return value
}
\description{
Unlock all project environment variablers
}
\examples{
proj.env = unlock(proj.env)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
