% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{unlock_proj}
\alias{unlock_proj}
\title{Unlock all the project variables}
\usage{
unlock_proj()
}
\value{
No return value
}
\description{
Unlock all project environment variablers
}
\examples{
unlock_proj()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
