% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{update_RDevVersion}
\alias{update_RDevVersion}
\title{Update stored R Development Version}
\usage{
update_RDevVersion()
}
\value{
No return value
}
\description{
Overwrites the stored R Version with the current R Version in .projectmaproot
}
\examples{
update_RDevVersion()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
