% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{update_req_packages}
\alias{update_req_packages}
\title{Update stored required packages}
\usage{
update_req_packages()
}
\value{
No return value
}
\description{
Overwrites the stored required packages with the current versions of installed packages
}
\examples{
update_req_packages()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
