% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{exit_proj}
\alias{exit_proj}
\title{Exit a project}
\usage{
exit_proj(reset_lib = T)
}
\arguments{
\item{reset_lib}{Boolean (T, F) indicator to reset the library path to its original state}
}
\value{
No return value
}
\description{
Exits a project by detaching the projectmap package and resetting the library path
}
\examples{
exit_proj()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
