% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{get_proj_root}
\alias{get_proj_root}
\title{Get the project root directory}
\usage{
get_proj_root(app = F)
}
\arguments{
\item{app}{Boolean (T, F) indicator to tell the function that it is being executed from within the app directory}
}
\value{
No return value
}
\description{
Updates the project environment with the root and current directories
}
\examples{
get_proj_root()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
