% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{library}
\alias{library}
\title{Redefined library, require, and install.packages functions to only look in the project library
This overwrites the base library function to only look in the user's project library to load a package}
\usage{
library(..., lib.loc = proj.env$libPath)
}
\arguments{
\item{...}{First, package name, then other named arguments passed to the base library function}

\item{lib.loc}{Path to users package library}
}
\value{
Normally library returns (invisibly) the list of attached packages, but TRUE or FALSE if logical.return is TRUE.
When called as library() it returns an object of class "libraryIQR", and for library(help=), one of class "packageInfo".
}
\description{
for package version control
}
\examples{
library(projectmap)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
