% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{load.packages}
\alias{load.packages}
\title{This function loads packages}
\usage{
load.packages(..., packages = NULL, character.only = F)
}
\arguments{
\item{pkgs}{Character vector of package names}
}
\value{
No return value
}
\description{
This function loads packages
}
\examples{
load.packages("packageName")
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
