% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{lock_proj}
\alias{lock_proj}
\title{Lock all the project variables}
\usage{
lock_proj()
}
\value{
No return value
}
\description{
Lock all project environment variablers
}
\examples{
lock_proj()
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
