% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{unlock}
\alias{unlock}
\title{Unlock the project environment}
\usage{
unlock(proj.env)
}
\value{
No return value
}
\description{
Resets the project environment to its default values.
}
\examples{
unlock(proj.env)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
